20.02.2022

+++ KonuYorumCore/DataAccess/Yorum.cs
@@ -8,7 +8,7 @@ namespace KonuYorumCore.DataAccess
         public int Id { get; set; }
         public string Icerik { get; set; }
         public string Yorumcu { get; set; }
-        public byte? Puan { get; set; }
+        public int? Puan { get; set; }
         public int KonuId { get; set; }
 
         public virtual Konu Konu { get; set; }
